+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.7.16
 
 * extended `Proxy`
+* made `Immutable` utilize `__slots__`