+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.7.17
+
+* bugfix for `satella.processes.call_and_return_stdout`
+ regarding encoding handling