+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.7.18
+
+* reformatted `ProcessFailed` exception
+* added an option not to check the return code